restructured to mvc
