+++ Quizzler-iOS13/Controller/ViewController.swift
@@ -1,0 +1,49 @@
+//
+//  ViewController.swift
+//  Quizzler-iOS13
+//
+//  Created by Angela Yu on 12/07/2019.
+//  Copyright © 2019 The App Brewery. All rights reserved.
+//
+
+import UIKit
+import AVFoundation
+
+class ViewController: UIViewController {
+    
+    @IBOutlet weak var questionLabel: UILabel!
+    @IBOutlet weak var trueButton: UIButton!
+    @IBOutlet weak var falseButton: UIButton!
+    @IBOutlet weak var progressBar: UIProgressView!
+    
+    var quizBrain = QuizBrain()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        updateUI()
+    }
+    
+    @IBAction func answerButtonPressed(_ sender: UIButton) {
+        let userAnswer: String = sender.currentTitle!
+        
+        if quizBrain.checkAnswer(userAnswer) {
+            sender.backgroundColor = UIColor.green
+        } else {
+            sender.backgroundColor = UIColor.red
+        }
+        
+        quizBrain.nextQuestion()
+        
+        Timer.scheduledTimer(timeInterval: 0.2, target: self, selector: #selector(updateUI), userInfo: nil, repeats: false)
+    }
+    
+    @objc func updateUI() {
+        questionLabel.text = quizBrain.getQuestionText()
+        progressBar.progress = quizBrain.getProgress()
+
+        trueButton.backgroundColor = UIColor.clear
+        falseButton.backgroundColor = UIColor.clear
+    }
+    
+}
+
