+++ Quizzler-iOS13/Model/QuizBrain.swift
@@ -1,21 +1,15 @@
 //
-//  ViewController.swift
+//  QuizBrain.swift
 //  Quizzler-iOS13
 //
-//  Created by Angela Yu on 12/07/2019.
-//  Copyright © 2019 The App Brewery. All rights reserved.
+//  Created by Leole Nkhatho on 2021/12/18.
+//  Copyright © 2021 The App Brewery. All rights reserved.
 //
 
-import UIKit
-import AVFoundation
+import Foundation
 
-class ViewController: UIViewController {
-    
-    @IBOutlet weak var questionLabel: UILabel!
-    @IBOutlet weak var trueButton: UIButton!
-    @IBOutlet weak var falseButton: UIButton!
-    @IBOutlet weak var progressBar: UIProgressView!
-    
+struct QuizBrain {
+    var questionNumber = 0
     let quiz = [
         Question(q: "A slug's blood is green.", a: "True"),
         Question(q: "Approximately one quarter of human bones are in the feet.", a: "True"),
@@ -31,40 +25,30 @@ class ViewController: UIViewController {
         Question(q: "Chocolate affects a dog's heart and nervous system; a few ounces are enough to kill a small dog.", a: "True")
     ]
     
-    var questionNumber = 0
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        updateUI()
-    }
-    
-    @IBAction func answerButtonPressed(_ sender: UIButton) {
-        let userAnswer: String = sender.currentTitle!
-        let actualAnswer: String = quiz[questionNumber].a
-        
-        if userAnswer == actualAnswer {
+    func checkAnswer(_ userAnswer: String) -> Bool {
+        if userAnswer == quiz[questionNumber].a {
             print("Correct")
-            sender.backgroundColor = UIColor.green
+            return true
         } else {
             print("Incorrect")
-            sender.backgroundColor = UIColor.red
+            return false
         }
-        
-        if questionNumber == quiz.count - 1 {
+    }
+    
+    func getQuestionText() -> String {
+        return quiz[questionNumber].q
+    }
+    
+    func getProgress() -> Float {
+
+        return Float(questionNumber + 1) / Float(quiz.count)
+    }
+    
+    mutating func nextQuestion() {
+        if questionNumber == quiz.count {
             questionNumber = 0
         } else {
             questionNumber += 1
         }
-        
-        Timer.scheduledTimer(timeInterval: 0.2, target: self, selector: #selector(updateUI), userInfo: nil, repeats: false)
     }
-    
-    @objc func updateUI() {
-        questionLabel.text = quiz[questionNumber].q
-        trueButton.backgroundColor = UIColor.clear
-        falseButton.backgroundColor = UIColor.clear
-        progressBar.progress = Float(questionNumber + 1) / Float(quiz.count)
-    }
-    
 }
-
